+++ YukoBot/Commands/Attributes/RequireRegisteredAttribute.cs
@@ -1,7 +1,6 @@
 ﻿using DSharpPlus.CommandsNext;
 using DSharpPlus.CommandsNext.Attributes;
 using System;
-using System.Linq;
 using System.Threading.Tasks;
 using YukoBot.Models.Database;
 using YukoBot.Models.Database.Entities;
@@ -9,19 +8,14 @@ using YukoBot.Models.Database.Entities;
 namespace YukoBot.Commands.Attributes
 {
     [AttributeUsage(AttributeTargets.Class | AttributeTargets.Method, AllowMultiple = false, Inherited = false)]
-    public sealed class RequireRegisteredAndNoBanAttribute : CheckBaseAttribute
+    public class RequireRegisteredAttribute : CheckBaseAttribute
     {
         public override Task<bool> ExecuteCheckAsync(CommandContext ctx, bool help)
         {
             ulong userId = ctx.Member.Id;
             YukoDbContext dbContext = new YukoDbContext();
             DbUser dbUser = dbContext.Users.Find(userId);
-            if (dbUser != null)
-            {
-                DbBan dbBan = dbContext.Bans.Where(x => x.UserId == userId && x.ServerId == ctx.Guild.Id).FirstOrDefault();
-                return Task.FromResult(dbBan == null);
-            }
-            return Task.FromResult(false);
+            return Task.FromResult(dbUser != null);
         }
     }
 }
